Prepare data for analysis

This function aggregates grocery store counts and computes additional variables based on raw census data. The function does not require any arguments but does require internet access.

Usage

prepare_data()

Format

A sf dataframe with 49 columns, the additional columns compared to the raw grocery_DA:

hsr_level

The quantile level transit stop number in that census DA (high, mid, or low)

n_hsr_stops

The transit stop number in that census DA

n_hsr_stops_density

The transit stop density per square kilometres in that census DA

pop_density

Population density per square kilometres

log_pop_density

Natural log of population density per square kilometres plus 1

log_area

Natural log of land area in square kilometres

dist_to_downtown

Euclidean distance from census DA centroid to Hamilton downton in metres

log_dist_to_downtown

Natural log of Euclidean distance from census DA centroid to Hamilton downton in metres

PCT_aged_under_24

Percentage of residents aged under 24 years old

PCT_aged_above_65

Percentage of residents aged above 65 years old

PCT_single_detached

Percentage of residents living in single detached houses

PCT_married_common_law

Percentage of residents married or living in common law

PCT_income_less_40k

Percentage of residents have household income less than 40k

PCT_income_greater_100k

Percentage of residents have household income greater than 100k

PCT_dont_know_official_language

Percentage of residents don't know/speak official languages

PCT_not_speak_offcial_language_at_home

Percentage of residents don't speak official languages at home

Value

a data frame based on grocery_DA with additional variables

Author

Zehui Yin, yinz39@mcmaster.ca

Examples

grocery_DA <- prepare_data()
#> Warning: attribute variables are assumed to be spatially constant throughout all geometries
#> Warning: st_centroid assumes attributes are constant over geometries
summary(grocery_DA)
#>     GeoUID          v_CA21_1..Population..2021
#>  Length:891         Min.   :   0              
#>  Class :character   1st Qu.: 441              
#>  Mode  :character   Median : 521              
#>                     Mean   : 639              
#>                     3rd Qu.: 664              
#>                     Max.   :7631              
#>                                               
#>  v_CA21_7..Land.area.in.square.kilometres v_CA21_8..Total...Age
#>  Min.   : 0.010                           Min.   :   0.0       
#>  1st Qu.: 0.090                           1st Qu.: 440.0       
#>  Median : 0.140                           Median : 520.0       
#>  Mean   : 1.255                           Mean   : 640.4       
#>  3rd Qu.: 0.310                           3rd Qu.: 665.0       
#>  Max.   :32.170                           Max.   :7630.0       
#>                                           NA's   :2            
#>  v_CA21_11..0.to.14.years v_CA21_10..Total...Age v_CA21_251..65.years.and.over
#>  Min.   :   0.0           Min.   :   0.0         Min.   :   0.0               
#>  1st Qu.:  60.0           1st Qu.: 220.0         1st Qu.:  70.0               
#>  Median :  75.0           Median : 265.0         Median :  90.0               
#>  Mean   : 102.7           Mean   : 325.8         Mean   : 117.3               
#>  3rd Qu.: 105.0           3rd Qu.: 335.0         3rd Qu.: 130.0               
#>  Max.   :1955.0           Max.   :3900.0         Max.   :1520.0               
#>  NA's   :2                NA's   :2              NA's   :2                    
#>  v_CA21_71..15.to.19.years v_CA21_89..20.to.24.years
#>  Min.   :  0.00            Min.   :  0.00           
#>  1st Qu.: 20.00            1st Qu.: 25.00           
#>  Median : 25.00            Median : 30.00           
#>  Mean   : 35.78            Mean   : 40.87           
#>  3rd Qu.: 40.00            3rd Qu.: 45.00           
#>  Max.   :570.00            Max.   :460.00           
#>  NA's   :2                 NA's   :2                
#>  v_CA21_435..Single.detached.house
#>  Min.   :   0.0                   
#>  1st Qu.:  80.0                   
#>  Median : 135.0                   
#>  Mean   : 140.7                   
#>  3rd Qu.: 180.0                   
#>  Max.   :1270.0                   
#>  NA's   :2                        
#>  v_CA21_434..Occupied.private.dwellings.by.structural.type.of.dwelling.data
#>  Min.   :   0.0                                                            
#>  1st Qu.: 170.0                                                            
#>  Median : 205.0                                                            
#>  Mean   : 250.5                                                            
#>  3rd Qu.: 270.0                                                            
#>  Max.   :2205.0                                                            
#>  NA's   :2                                                                 
#>  v_CA21_453..Marital.status.for.the.total.population.aged.15.years.and.over
#>  Min.   :   0.0                                                            
#>  1st Qu.: 375.0                                                            
#>  Median : 440.0                                                            
#>  Mean   : 537.7                                                            
#>  3rd Qu.: 570.0                                                            
#>  Max.   :5680.0                                                            
#>  NA's   :2                                                                 
#>  v_CA21_456..Married.or.living.common.law
#>  Min.   :   0.0                          
#>  1st Qu.: 190.0                          
#>  Median : 235.0                          
#>  Mean   : 293.2                          
#>  3rd Qu.: 310.0                          
#>  Max.   :3605.0                          
#>  NA's   :2                               
#>  v_CA21_1142..Gini.index.on.adjusted.household.after.tax.income
#>  Min.   :0.1820                                                
#>  1st Qu.:0.2260                                                
#>  Median :0.2475                                                
#>  Mean   :0.2558                                                
#>  3rd Qu.:0.2770                                                
#>  Max.   :0.9120                                                
#>  NA's   :15                                                    
#>  v_CA21_1144..Knowledge.of.official.languages.for.the.total.population.excluding.institutional.residents
#>  Min.   :   0.0                                                                                         
#>  1st Qu.: 440.0                                                                                         
#>  Median : 515.0                                                                                         
#>  Mean   : 634.2                                                                                         
#>  3rd Qu.: 655.0                                                                                         
#>  Max.   :7630.0                                                                                         
#>  NA's   :2                                                                                              
#>  v_CA21_671..With.total.income v_CA21_674..Under..10.000..including.loss.
#>  Min.   : 155.0                Min.   :  5.00                            
#>  1st Qu.: 360.0                1st Qu.: 25.00                            
#>  Median : 420.0                Median : 35.00                            
#>  Mean   : 512.1                Mean   : 41.74                            
#>  3rd Qu.: 535.0                3rd Qu.: 45.00                            
#>  Max.   :5360.0                Max.   :545.00                            
#>  NA's   :15                    NA's   :15                                
#>  v_CA21_677...10.000.to..19.999 v_CA21_680...20.000.to..29.999
#>  Min.   : 10.00                 Min.   : 20.00                
#>  1st Qu.: 40.00                 1st Qu.: 50.00                
#>  Median : 55.00                 Median : 70.00                
#>  Mean   : 63.33                 Mean   : 81.44                
#>  3rd Qu.: 70.00                 3rd Qu.: 90.00                
#>  Max.   :595.00                 Max.   :725.00                
#>  NA's   :15                     NA's   :15                    
#>  v_CA21_683...30.000.to..39.999 v_CA21_686...40.000.to..49.999
#>  Min.   : 10.00                 Min.   : 10.0                 
#>  1st Qu.: 45.00                 1st Qu.: 40.0                 
#>  Median : 55.00                 Median : 50.0                 
#>  Mean   : 66.32                 Mean   : 57.6                 
#>  3rd Qu.: 75.00                 3rd Qu.: 65.0                 
#>  Max.   :635.00                 Max.   :530.0                 
#>  NA's   :15                     NA's   :15                    
#>  v_CA21_689...50.000.to..59.999 v_CA21_692...60.000.to..69.999
#>  Min.   :  0.00                 Min.   :  0.00                
#>  1st Qu.: 30.00                 1st Qu.: 25.00                
#>  Median : 40.00                 Median : 30.00                
#>  Mean   : 46.54                 Mean   : 37.14                
#>  3rd Qu.: 50.00                 3rd Qu.: 40.00                
#>  Max.   :470.00                 Max.   :395.00                
#>  NA's   :15                     NA's   :15                    
#>  v_CA21_695...70.000.to..79.999 v_CA21_698...80.000.to..89.999
#>  Min.   :  0.00                 Min.   :  0.00                
#>  1st Qu.: 15.00                 1st Qu.: 10.00                
#>  Median : 25.00                 Median : 15.00                
#>  Mean   : 28.56                 Mean   : 21.38                
#>  3rd Qu.: 30.00                 3rd Qu.: 25.00                
#>  Max.   :335.00                 Max.   :260.00                
#>  NA's   :15                     NA's   :15                    
#>  v_CA21_701...90.000.to..99.999 v_CA21_707...100.000.to..149.999
#>  Min.   :  0.0                  Min.   :  0.00                  
#>  1st Qu.: 10.0                  1st Qu.: 15.00                  
#>  Median : 15.0                  Median : 25.00                  
#>  Mean   : 16.8                  Mean   : 35.88                  
#>  3rd Qu.: 20.0                  3rd Qu.: 45.00                  
#>  Max.   :235.0                  Max.   :505.00                  
#>  NA's   :15                     NA's   :15                      
#>  v_CA21_710...150.000.and.over v_CA21_1156..Neither.English.nor.French
#>  Min.   :  0.00                Min.   :  0.00                         
#>  1st Qu.:  0.00                1st Qu.:  0.00                         
#>  Median :  5.00                Median :  5.00                         
#>  Mean   : 14.95                Mean   : 11.57                         
#>  3rd Qu.: 20.00                3rd Qu.: 15.00                         
#>  Max.   :165.00                Max.   :195.00                         
#>  NA's   :15                    NA's   :2                              
#>  v_CA21_2176..Non.official.language
#>  Min.   :  0.00                    
#>  1st Qu.: 10.00                    
#>  Median : 25.00                    
#>  Mean   : 40.05                    
#>  3rd Qu.: 50.00                    
#>  Max.   :810.00                    
#>  NA's   :2                         
#>  v_CA21_2167..All.languages.spoken.at.home.for.the.total.population.excluding.institutional.residents
#>  Min.   :   0.0                                                                                      
#>  1st Qu.: 440.0                                                                                      
#>  Median : 515.0                                                                                      
#>  Mean   : 634.2                                                                                      
#>  3rd Qu.: 655.0                                                                                      
#>  Max.   :7630.0                                                                                      
#>  NA's   :2                                                                                           
#>       Freq          n_hsr_stops              geometry   n_hsr_stops_density
#>  Min.   :0.00000   Min.   : 0.000   MULTIPOLYGON :891   Min.   :  0.00     
#>  1st Qu.:0.00000   1st Qu.: 0.000   epsg:26917   :  0   1st Qu.:  0.00     
#>  Median :0.00000   Median : 2.000   +proj=utm ...:  0   Median : 11.76     
#>  Mean   :0.09877   Mean   : 2.585                       Mean   : 15.04     
#>  3rd Qu.:0.00000   3rd Qu.: 4.000                       3rd Qu.: 22.22     
#>  Max.   :4.00000   Max.   :49.000                       Max.   :200.00     
#>                                                                            
#>  hsr_level   pop_density    log_pop_density     log_area     
#>  low :450   Min.   :    0   Min.   : 0.000   Min.   :-4.605  
#>  mid :220   1st Qu.: 2068   1st Qu.: 7.635   1st Qu.:-2.408  
#>  high:221   Median : 3655   Median : 8.204   Median :-1.966  
#>             Mean   : 4538   Mean   : 7.835   Mean   :-1.515  
#>             3rd Qu.: 5352   3rd Qu.: 8.585   3rd Qu.:-1.171  
#>             Max.   :79800   Max.   :11.287   Max.   : 3.471  
#>                                                              
#>  dist_to_downtown.V1 log_dist_to_downtown.V1 PCT_aged_under_24
#>  Min.   :  159.377   Min.   : 5.071275       Min.   : 0.00    
#>  1st Qu.: 3603.999   1st Qu.: 8.189794       1st Qu.:18.52    
#>  Median : 6012.537   Median : 8.701602       Median :21.43    
#>  Mean   : 6901.449   Mean   : 8.582862       Mean   :21.34    
#>  3rd Qu.: 8962.513   3rd Qu.: 9.100806       3rd Qu.:24.28    
#>  Max.   :28288.735   Max.   :10.250219       Max.   :46.77    
#>                                              NA's   :3        
#>  PCT_aged_above_65 PCT_single_detached PCT_married_common_law
#>  Min.   : 0.00     Min.   :  0.00      Min.   :16.98         
#>  1st Qu.:13.10     1st Qu.: 37.86      1st Qu.:46.74         
#>  Median :17.50     Median : 72.44      Median :55.23         
#>  Mean   :19.37     Mean   : 63.79      Mean   :54.05         
#>  3rd Qu.:23.08     3rd Qu.: 93.94      3rd Qu.:61.73         
#>  Max.   :94.67     Max.   :104.17      Max.   :73.68         
#>  NA's   :3         NA's   :4           NA's   :3             
#>  PCT_income_less_40k PCT_income_greater_100k PCT_dont_know_official_language
#>  Min.   :30.43       Min.   : 0.000          Min.   : 0.000                 
#>  1st Qu.:43.18       1st Qu.: 3.792          1st Qu.: 0.000                 
#>  Median :48.69       Median : 7.380          Median : 1.205                 
#>  Mean   :49.81       Mean   : 9.356          Mean   : 1.693                 
#>  3rd Qu.:55.28       3rd Qu.:13.711          3rd Qu.: 2.381                 
#>  Max.   :92.16       Max.   :37.500          Max.   :17.241                 
#>  NA's   :15          NA's   :15              NA's   :3                      
#>  PCT_not_speak_offcial_language_at_home
#>  Min.   : 0.000                        
#>  1st Qu.: 2.222                        
#>  Median : 5.093                        
#>  Mean   : 5.775                        
#>  3rd Qu.: 8.266                        
#>  Max.   :29.474                        
#>  NA's   :3